backend/pyproject.toml
[tool.poetry]
name = "remoteservermanager"
version = "0.1.0"
description = ""
authors = ["vovquewa <kit.jke@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.32.1"
paramiko = "^3.5.0"
pydantic = "^2.10.2"
pydantic-settings = "^2.6.1"
alembic = "^1.14.0"
aiosqlite = "^0.20.0"
python-dotenv = "^1.0.1"
fastapi = "^0.115.5"
fastapi-users-db-sqlalchemy = "^6.0.1"
asyncpg = "^0.30.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
